Phrecon.pl -- Added deprecation message, changed output to 60-char lines
Version 2.2 is intended to be the last version of the perl-version of Phrecon.
Future development will be done on the python version (phrecon.py) but the perl
script is kept here in case it is needed for some reason.

Final changes include a deprecation message, 60-char FASTA output lines (instead of 70),
and removing the extraneous newline between the ref base sequence line and the first
snp sequence record.