v4.4.0 - Fix sorting error introduced by chrom+pos handling
This may be a temporary fix but at least sorts loci properly again
(numeric vs lexical) for non-multichromosome VCF snp file inputs.  For
multichromosome VCF snp file inputs, it grabs the trailing numeral
sequence from the loci tag and treats it as the real numeric position
(this may need to revised later if the assumption is false).